This was an epic adventure! 🤘
So, this repository pretty much reached its limits, and it's time to
move on 🚀. Everything started 3 years ago when I tried to configure a
jenkins server with unity to build some games when I commit. Then I
asked myself: How can we make it easier to build and test unity
projects in the cloud? I think this repository was proof that it is
possible to solve this problem by combining our efforts and make it
more accessible. Hopefully, the repository was helpful to quite a few
ones here. Thanks to everyone who contributed in a way or another to
this project. It is not the end and existing images will stay on docker
hub (unless they delete them, you know).

It will get better! We now have an automated system to build and
release Unity docker images, we have a versioning system enabling us to
publish new images without breaking previous builds. We have a better
and safer way to install unity components using unity hub and even
better, we have a community ❤.

ReadMe says it all, moving to https://github.com/game-ci/docker/ is the
right step forward. See you on discord at https://game.ci/discord
🎮

Many thanks to everyone who contributed to gableroux/unity3d and
special thanks to

* Webber
* Fish
* David Finol
* Théo Sabattié
* Szymon Kazmierczak
* Dharmesh Prajapati
* kalle (jag)
* Mathieu Tortuyaux
* Devashish Lal (BLaZeKiLL)
* Jonathan Perret